[GBL-19660] Report hitting the max concurrency limit to the task runner
This event is reported when the limit is hit the first time after a
task runner is called. One may say the limit is hit again at every
tick in the load generator until a task process eventually terminates,
but it doesn't make sense to report this event every ms for the task
runner until that time.