[GBL-21180] Avoid the load generator failing behind schedule
The load generator keeps track with sub-ms accuracy the next time a
task needs to be triggered. This value is updated based on the
intensity only, independent of the wall clock time.

The generator process is scheduled on the wall clock's ticks however,
and therefore will always be slightly behind its intended
schedule. When the generator may not trigger more than once per tick,
it has no way of recovering from accumulated delay.

Consider for example the scheduler shall wake up at `Ti` to trigger
the i-th task, but for actually gets woken up at `T > Ti` (this is
perfectly normal with Erlang's scheduling algorithms). After
triggering the i-th task, the trigger time for the i+1-th is
calculated as `T(i+1)`. At this point it may happen that `T > T(i+1)`,
so the correct strategy would be to immediately trigger that task too,
not to wait one more tick.

As a side effect, this strategy would allow modeling loads above 500Hz
too (although the exact replication of such high frequency,
non-constant waveforms may not be possible).

One more trick helps improving the reproduction of non-constant
waveforms: when the load specification is polled after triggering a
task, the trigger time shall be used as the current time, not the
elapsed time derived from the wall clock.

Finally, since the trigger time is already represented as a float,
there is no reason to round the elapsed time to whole
milliseconds. That would only reduce the accurracy of the
scheduling. Truncating the elapsed time to whole milliseconds is only
necessary before passing the value to the load specification function
(since it takes an integer argument).