introduce dynamic configuration of max_concurrent
Previously, the only way to alter the max_concurrent option (which
limits the number of ongoing tasks; spawned processes a load generator
may have), was to remove the load generator and re-add it with a
different setting.

This option is useful for protecting a load slave from overloading. It
can also be used to simulate a traditional worker based load generation.